forgot to make ictop and icbot floats
Former-commit-id: 3ea84fabfface1ff7cb2ee4ad482491f8438e72e

+++ src/interpol_mod.f90
@@ -1267,7 +1267,7 @@ subroutine interpol_rain(itime,kz,yint1,yint2,yint3,ytint,yint4,intiy1,intiy2,ic
   implicit none
 
   integer, intent(in) :: kz, icmv, itime
-  integer, intent(out) :: intiy1,intiy2
+  real, intent(out) :: intiy1,intiy2
   real, intent(out) :: yint1,yint2,yint3,ytint,yint4
   integer :: m
   integer :: mm,ip1,ip2,ip3,ip4
